# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Discount Service API",
    "version": "1.0.0",
    "description": "API for managing vouchers, promotions, and applying discounts to orders",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "schemas": {
...
      "Promotion": {
        "type": "object",
        "required": [
          "discountType",
          "discountValue",
          "expirationDate",
          "usageLimit"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Promotion code (auto-generated if not provided)",
            "example": "SUMMER2024"
          },
          "eligibleProductCategories": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Product categories eligible for this promotion",
            "example": [
              "electronics",
              "clothing"
            ]
          },
          "eligibleProductIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Specific product IDs eligible for this promotion",
            "example": [
              "prod1",
              "prod2"
            ]
          },
          "discountType": {
            "type": "string",
            "enum": [
              "percentage",
              "fixed"
            ],
            "description": "Type of discount",
            "example": "percentage"
          },
          "discountValue": {
            "type": "number",
            "description": "Discount value (percentage or fixed amount)",
            "example": 15
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time",
            "description": "Expiration date",
            "example": "2024-12-31T23:59:59Z"
          },
          "usageLimit": {
            "type": "integer",
            "description": "Maximum number of times the promotion can be used",
            "example": 500
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the promotion is active",
            "example": true
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Validation error details"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/promotions": {
      "get": {
        "summary": "Get all promotions",
        "tags": [
          "Promotions"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by active status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of promotions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Promotion"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new promotion",
        "tags": [
          "Promotions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Promotion created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Promotion"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Promotion code already exists"
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      }
    },
    "/api/promotions/{id}": {
      "get": {
        "summary": "Get promotion by ID",
        "tags": [
          "Promotions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Promotion ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Promotion details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Promotion"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Promotion not found"
          }
        }
      },
      "put": {
        "summary": "Update a promotion",
        "tags": [
          "Promotions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Promotion ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "eligibleProductCategories": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "eligibleProductIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "discountType": {
                    "type": "string",
                    "enum": [
                      "percentage",
                      "fixed"
                    ]
                  },
                  "discountValue": {
                    "type": "number"
                  },
                  "expirationDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "usageLimit": {
                    "type": "integer"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Promotion updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Promotion"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Promotion not found"
          },
          "409": {
            "description": "Promotion code already exists"
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a promotion",
        "tags": [
          "Promotions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Promotion ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Promotion deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Promotion not found"
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}